backend/pyproject.toml
[project]
name = "yo-agent-backend"
version = "0.1.0"
description = "YoAgent Backend API"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.13",
    "pydantic-settings>=2.0",
    "redis>=5.0",
    "psycopg2-binary>=2.9.11",
    "langgraph>=1.1.2",
    "langchain-openai>=1.1.11",
    "langgraph-checkpoint-postgres>=3.0.4",
    "psycopg[binary]>=3.1",
    "python-slugify>=8.0.4",
]
